+++ packages/georedus-ui/src/GeoReDUS/GeoReDUS.jsx
@@ -705,8 +705,7 @@ function GeoReDUSInner({
         }}
         direction="row"
         gap="4"
-        alignItems="center"
-      >
+        alignItems="center">
         <ViewLayoutPopover
           viewSpecs={viewSpecsQuery.data}
           viewConfState={viewConfState}
@@ -718,8 +717,7 @@ function GeoReDUSInner({
           width="400px"
           direction="column"
           maxWidth="30vw"
-          gap="3"
-        >
+          gap="3">
           <HotfixSelectLargeFont>
             <Input
               schema={MUNICIPIO_ID_SELECTOR_SCHEMA}
@@ -732,8 +730,7 @@ function GeoReDUSInner({
               alignSelf: 'flex-end',
               position: 'absolute',
               top: '100%',
-            }}
-          >
+            }}>
             <Input
               schema={{
                 type: 'booleanCheckbox',
@@ -840,14 +837,12 @@ function GeoReDUSInner({
                   boxShadow: 'none',
                   opacity: legends.length > 0 ? 1 : 0,
                 }}
-                position="bottom-right"
-              >
+                position="bottom-right">
                 {legends.length > 0 && (
                   <LegendContainer
                     direction="row"
                     gap="3"
-                    p={resolvedLayout.length > 1 ? '3' : '4'}
-                  >
+                    p={resolvedLayout.length > 1 ? '3' : '4'}>
                     {resolvedLayout.length > 1 && (
                       <Tooltip content="Fechar visualização">
                         <IconButton
@@ -858,8 +853,7 @@ function GeoReDUSInner({
                               type: 'DEACTIVATE_VIEW',
                               payload: views[0].id,
                             })
-                          }
-                        >
+                          }>
                           <Icon path={mdiClose} size="20px" />
                         </IconButton>
                       </Tooltip>
@@ -868,18 +862,16 @@ function GeoReDUSInner({
                     <EvenSpacedList
                       columns={legends.length > 1 ? 2 : 1}
                       gap="10px"
-                    >
+                      style={{ maxWidth: '300px' }}>
                       {legends.map((legend) => (
                         <HoverLegend
                           {...(resolvedLayout.length > 1
                             ? {
                                 direction: 'row',
-                                maxWidth: '300px',
                                 size: '1',
                               }
                             : {
                                 direction: 'column',
-                                maxWidth: '150px',
                                 size: '2',
                               })}
                           key={legend.id}
@@ -904,8 +896,7 @@ function GeoReDUSInner({
                       width: 100,
                       height: 100,
                       boxShadow: 'none',
-                    }}
-                  >
+                    }}>
                     <MapStyleToggleCtrl
                       style={{
                         position: 'relative',
@@ -917,8 +908,7 @@ function GeoReDUSInner({
                         setBaseMapStyle(
                           baseMapStyle === 'dataviz' ? 'satellite' : 'dataviz',
                         )
-                      }
-                    >
+                      }>
                       <MapWindow
                         style={{
                           pointerEvents: 'none',
@@ -948,8 +938,7 @@ function GeoReDUSInner({
               {mapProps.children || null}
             </>
           ),
-        }))}
-      >
+        }))}>
         {(viewsLoading || tilesLoading) && (
           <LoadingIndicator
             style={{