+++ packages/georedus-ui/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orioro/template-react
 
+## 0.19.5
+
+### Patch Changes
+
+- fix custom geojson aggregation
+
+## 0.19.4
+
+### Patch Changes
+
+- update dependencies
+
+## 0.19.3
+
+### Patch Changes
+
+- fix total_pessoas_por_km2 mainDisplayValue
+- update dependency versions
+- Updated dependencies
+  - @orioro/react-maplibre-util@0.7.1
+  - @orioro/react-chart-util@0.3.1
+  - @orioro/vector-tile-util@0.3.1
+  - @orioro/react-dir-nav@0.2.1
+
 ## 0.19.2
 
 ### Patch Changes